pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot"
version = "0.1.0"
description = "Gemini customer chatbot with RAG and WhatsApp"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "sqlalchemy>=2.0.36",
    "typer>=0.12.0",
    "streamlit>=1.39.0",
    "pypdf>=5.0.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "lancedb>=0.14.0",
    "numpy>=1.26.0",
    "fasttext>=0.9.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
